AndroidTestRunner: print test results to stdout as they go
The testrunner was always printing the results only at the end of the
test run. If tests are long, it means waiting long to have any sort of
result. With this, the results are printed to stdout whenever they are
written by qtest on the device.

Fixes: QTBUG-129975
Change-Id: If4315eb74e73f6274a735e56be7e6989563bb514
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>
(cherry picked from commit 7545e7c921f364cdb037f8e967f012351c2eb3ad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>