Windows: Fix missing glyphs for very large font
When the font size gets to a certain size (at some point
after 1200 pixels), DirectWrite will just return an empty
rect as the bound for it and does not want to render it. At
this point, you might want to consider other ways of
rendering the text, but since the same did not happen
e.g. with the Freetype backend, it was an unfortunate
platform difference.

As a fail safe, we detect when this occurs and fall back to
the QFontEngine base class approach of drawing the glyph as a
QPainterPath instead.

[ChangeLog][Windows] Fixed an issue where glyphs might be missing
for very large fonts.

Fixes: QTBUG-126671
Change-Id: I13cc1e5d71fae114e2a7933a7c5f7b9980237390
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit a5ddcbb76b674470f0c3d1d5e0bb7697a2828d25)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>